Fixed build issues
The problem with the linker was being caused by the Makefile.  I could not figure out why it was failing, so I rewrote the Makefile from scratch.
I also needed to comment out some code due to incompatibility between the c code that was written prior to the ANSII standard.  This mostly had to do with the file processing functions, which will need to be rewritten using the present file functions.