Added comments to used5.asm.
These comments were in the documentation so I want them in my test case as well.

Please not that this file does not use tabs and everything is in lower case except were applicable in the comments.  help.asm uses tabs and every thing is in upper case.  This was intentional to test the assembler's capability to handle upper and lower case along with tabs.